requirements.txt
streamlit
langchain-google-genai
pymupdf
pandas
langchain
python-dotenv